+++ loopx/heartbeat_prompt.py
@@ -542,7 +542,6 @@ def build_heartbeat_prompt(
         "registered_agents": normalized_registered_agents,
         "primary_agent": normalized_primary_agent,
         "side_agent_handoff_agent": normalized_side_agent_handoff_agent,
-        "available_capabilities": normalized_available_capabilities,
         "expanded_prompt_command": expanded_prompt_command,
         "compact_prompt_command": compact_prompt_command,
         "brief_prompt_command": brief_prompt_command,
@@ -628,7 +627,6 @@ def build_heartbeat_prompt_error_payload(
         "registered_agents": normalized_registered_agents,
         "primary_agent": str(primary_agent).strip() if primary_agent else None,
         "side_agent_handoff_agent": str(side_agent_handoff_agent).strip() if side_agent_handoff_agent else None,
-        "available_capabilities": projected_available_capabilities,
         "expanded_prompt_command": expanded_prompt_command,
         "compact_prompt_command": compact_prompt_command,
         "brief_prompt_command": brief_prompt_command,
